Make forward.selection() work also if choose.from is a numeric vector.
When specifying the indices in choose.from, the code was failing when the
input was a number such as 5 or a vector such as c(5, 10), because in both
cases these are considered numeric.

This makes forward.selection() work also in those cases, as well as testing
for the presence of floating point values, which cannot be used as indices.