Remove unnecessary discarding NAs, since they cannot arise anymore.
This was originally meant to address the case when the initial model was
specified as a vector of strings and contained an interaction term. In such
case, the matching produced an NA for terms such as "age*sex", and the code
was removing them.

Now we have already checked that all variables mentioned in init.model are
available, and interaction terms are dealt with it correctly, so it should
not be possible for keep.vars to contain any NA.