Fix error when a categorical variable is selected.
The change in 0cae4fb broke the subsetting to extract the model coefficients
if factor variables were selected. The error was produced at this point in
nested.forward.selection():

   model$summary[panel, "Estimate"]

because for factor variables, the row names of model$summary also contain
the level label, so the original variable name listed in panel will not
match. As a solution, we now perform a partial match so that we can retrieve
the coefficient correctly if the factor had only two levels (the most common
case).

This will not work if the factor has 3 or more levels (as pmatch() returns
NA in this case): this means that instead of reporting a coefficient for
these variables, we will report NA. This is probably fine because reporting
just one of the many coefficients would be inappropriate anyway. Besides,
this additional column is an undocumented feature that is reported only for
convenience and is not used anywhere else in the code.