Add new interfaces based on formulas.
This adds the new functions fs() and nested.fs() and changes the interface
of nested.glm() so that they all use a formula interface similar to the one
of glm().

The older forward.selection() and nested.forward.selection() have been turned
into wrappers around the new interface for backward compatibility.